Resolves #119 - Combined all tests into one method
