--- docroot/profiles/custom/cgov_site/modules/custom/cgov_core/tests/src/Kernel/FieldStorage/NodeBrowserTitleFieldStorageTest.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace Drupal\Tests\cgov_core\Kernel\FieldStorage;
-
-/**
- * Tests node Browser Title field storage.
- *
- * @group cgov_core
- */
-class NodeBrowserTitleFieldStorageTest extends CGovFieldStorageTestBase {
-
-  /**
-   * Modules to enable.
-   *
-   * @var array
-   */
-  public static $modules = [
-    'user', 'system', 'file', 'field', 'node', 'text', 'filter', 'workflows', 'content_moderation',
-    'language', 'content_translation', 'cgov_core',
-  ];
-
-  /**
-   * Tests the Browser Title field.
-   */
-  public function testField() {
-    $this->baseTestPlainTextField('field_browser_title', 'Browser Title');
-  }
-
-}
